updating the list - and source
List so that we don’t see the .git folders.  We use git to manage Doc
folders, so .git folders are in among the Doc/.  This makes them not
appear.  Source now will take you to the Source (sorry for the
overloading of terms) of the text if the it does not render into a
document. (I.e. for lists of things, etc.)  Beats looking at a blank
page.  And  change from the French “Modifier”.